requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements.txt requirements.in
#
aiohttp==3.5.4
aioredis==1.2.0
async-timeout==3.0.1      # via aiohttp, aioredis
asyncpg==0.18.3
attrs==19.1.0             # via aiohttp
chardet==3.0.4            # via aiohttp
gitdb2==2.0.5             # via gitpython
gitpython==2.1.11
hiredis==1.0.0            # via aioredis
idna==2.8                 # via yarl
multidict==4.5.2          # via aiohttp, yarl
pyyaml==5.1
smmap2==2.0.5             # via gitdb2
yarl==1.3.0               # via aiohttp
